pyproject.toml
[tool.poetry]
name = "qfat"
version = "0.1.0"
description = ""
authors = ["Ziyad Sheebaelhamd <zsheebaelhamd@gmail.com>"]
readme = "README.md"
package-mode=true
packages = [{include = "qfat", from="src"}, {include = "hydra_plugins", from = "src"}] 

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pre-commit = "^3.8.0"
torch = "^2.4.1"
omegaconf = "^2.3.0"
numpy = "<=2.1.1"
hydra-core = "^1.3.2"
wandb = {extras = ["media"], version = ">0.18.0"}
einops = "^0.8.0"
matplotlib = "^3.9.2"
dm_control= "<=1.0.20" # franka kitchen compatibility
joblib = "^1.4.2"
gym = "^0.26.2"
notebook = "^7.2.2"
line-profiler = "^4.1.3"
plotly = "^5.24.1"
pymunk = "^6.9.0"
shapely = "^2.0.6"
pygame = "^2.6.1"
opencv-python = "^4.10.0.84"
scikit-image = "^0.24.0"
memory-profiler = "^0.61.0"
stable-baselines3 = "^2.3.2"
zarr = "^2.18.3"
glfw = "^2.7.0"
Cython = "<3.0.0"
gymnasium = {extras = ["mujoco"], version = "^1.0.0"}
torchvision = "^0.20.1"
pytest = "^8.3.4"
termcolor = "^2.5.0"
scikit-learn = "^1.6.1"
latex = "^0.7.0"
imageio = "^2.37.0"
imageio-ffmpeg = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
